# server/swagger-doc/swagger-api.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"register","summary":"Register a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully registered."},"400":{"description":"Invalid request data. Please provide valid user information."},"409":{"description":"A user with this username or email already exists"},"500":{"description":"Internal server error occurred. Please try again later."}},"tags":["Authentication"]}},"/auth/verifyEmail":{"post":{"operationId":"verifyEmail","summary":"Verify email verification code","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyCodeDto"}}}},"responses":{"200":{"description":"Email verification code verified successfully."},"404":{"description":"Not Found - User not found."},"406":{"description":"Not Acceptable - Verification code incorrect."},"500":{"description":"Internal Server Error - Failed to verify reset code. Please try again later."}},"tags":["Authentication"]}},"/auth/signIn":{"post":{"operationId":"signIn","summary":"Sign in to the system","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User signed in successfully."},"401":{"description":"Invalid credentials. Please check your email and password."},"409":{"description":"A user with this username or email already exists"},"500":{"description":"Internal server error occurred. Please try again later."}},"tags":["Authentication"]}},"/auth/forgotPassword":{"post":{"operationId":"forgotPassword","summary":"Initiate password reset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"Reset code verified successfully."},"404":{"description":"User not found."},"500":{"description":"Failed to send password reset code. Please try again later.d."}},"tags":["Authentication"]}},"/auth/verifyResetCode":{"post":{"operationId":"verifyResetCode","summary":"Verify password reset code","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyCodeDto"}}}},"responses":{"200":{"description":"Reset code verified successfully."},"404":{"description":"Not Found - User not found."},"406":{"description":"Not Acceptable - Verification code incorrect."},"500":{"description":"Internal Server Error - Failed to verify reset code. Please try again later."}},"tags":["Authentication"]}},"/auth/updatePassword":{"post":{"operationId":"updatePassword","summary":"Update user password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"200":{"description":"Password updated successfully"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"User not found"},"500":{"description":"Internal server error"}},"tags":["Authentication"]}},"/auth/resetPassword":{"post":{"operationId":"resetPassword","summary":"Reset password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Password reset successfully."},"404":{"description":"Not Found - User not found."},"500":{"description":"Internal Server Error - Failed to reset password. Please try again later."}},"tags":["Authentication"]}},"/users/me":{"get":{"operationId":"getMe","summary":"Fetch current user (own) profile","parameters":[],"responses":{"200":{"description":"Return the current user profile."},"401":{"description":"Unauthorized"},"404":{"description":"User profile cannot be found"},"500":{"description":"Internal Server Error - Failed to get user profile. Please try again later."}},"tags":["Users"]}},"/users/{username}":{"get":{"operationId":"getUserByUsername","summary":"Get user by username","parameters":[{"name":"username","required":true,"in":"path","description":"Username of the user to retrieve","schema":{"type":"string"}}],"responses":{"200":{"description":"User found and retrieved"},"404":{"description":"User not found"},"500":{"description":"Failed to find user"}},"tags":["Users"]}},"/users/me/update":{"patch":{"operationId":"update","summary":"Update current user (own) profile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Profile update successfully"},"400":{"description":"User provided invalid username and/or email"},"500":{"description":"The server is experiencing an error. "}},"tags":["Users"]}},"/users/me/education/update/{educationId}":{"patch":{"operationId":"updateEducation","summary":"Update user education information","parameters":[{"name":"educationId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EducationInfoDto"}}}},"responses":{"200":{"description":"Education updated successfully"},"404":{"description":"User or education not found"},"500":{"description":"Failed to update education"}},"tags":["Users"]}},"/users/me/education/remove/{educationId}":{"delete":{"operationId":"removeEducation","summary":"Delete user education information","parameters":[{"name":"educationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Education deleted successfully"},"404":{"description":"User or education not found"},"500":{"description":"Failed to delete education"}},"tags":["Users"]}},"/users/me/delete":{"delete":{"operationId":"delete","summary":"Delete current user (own) account","parameters":[],"responses":{"200":{"description":"Profile deleted successfully"},"404":{"description":"User not found"},"500":{"description":"The server is experiencing an error. "}},"tags":["Users"]}},"/users/me/skills/remove/{skillId}":{"patch":{"operationId":"removeSkill","summary":"Remove skill from current user (own) profile","parameters":[{"name":"skillId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Skill removed successfully"},"404":{"description":"User not found or skill not found"},"500":{"description":"The server is experiencing an error. "}},"tags":["Users"]}},"/locations":{"post":{"operationId":"create","summary":"Create a new location","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLocationDto"}}}},"responses":{"201":{"description":"Location created successfully."},"409":{"description":"Conflict - Location already exists."},"500":{"description":"Internal Server Error - Failed to create location."}},"tags":["Locations"]},"get":{"operationId":"findAll","summary":"Get all locations","parameters":[],"responses":{"200":{"description":"Locations found successfully."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/id/{id}":{"get":{"operationId":"findOneById","summary":"Get a location by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/name/{name}":{"get":{"operationId":"findOneByName","summary":"Get a location by name","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified name cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/code/{code}":{"get":{"operationId":"findOneByShortCode","summary":"Get a location by short code","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified short code cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/{id}":{"patch":{"operationId":"update","summary":"Update a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLocationDto"}}}},"responses":{"200":{"description":"Location updated successfully."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to update location. Please try again later."}},"tags":["Locations"]},"delete":{"operationId":"remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location successfully deleted."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to delete location. Please try again later."}},"tags":["Locations"]}},"/files/profilePic/update":{"post":{"operationId":"uploadFile","summary":"Update user profile picture","parameters":[{"name":"file","required":true,"in":"path","schema":{"format":"binary","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfilePicDto"}}}},"responses":{"201":{"description":"User profile picture updated successfully"},"401":{"description":"Unauthorized"},"404":{"description":"User with specified email does not exist"},"500":{"description":"Failed to update user profile image. Please try again later."}},"tags":["Files"]}},"/files/getProfilePicture/{email}":{"get":{"operationId":"serveProfilePicture","summary":"Get profile picture of a user by email","parameters":[{"name":"email","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Profile picture retrieved successfully"},"404":{"description":"User or profile picture not found"},"500":{"description":"Internal Server Error - Failed to fetch profile picture. Please try again later."}},"tags":["Files"]}},"/files/deleteProfilePicture":{"delete":{"operationId":"deleteAvatar","summary":"Delete own profile picture.","parameters":[],"responses":{"200":{"description":"Profile picture deleted successfully"},"404":{"description":"User with the specified id not found"},"500":{"description":"Internal Server Error - Failed to delete profile picture. Please try again later."}},"tags":["Files"]}},"/files/logo/update/{id}":{"post":{"operationId":"uploadLogo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Files"]}},"/files/getLogo/{id}":{"get":{"operationId":"serveLogo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Files"]}},"/files/deleteLogo/{id}":{"delete":{"operationId":"deleteLogo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Files"]}},"/files/uploadPermit/{id}":{"post":{"operationId":"uploadPermit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Files"]}},"/files/getPermit/{id}":{"get":{"operationId":"servePermit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Files"]}},"/skills":{"post":{"operationId":"create","summary":"Create a new skill","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSkillDto"}}}},"responses":{"201":{"description":"Skill created successfully"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Category not found"},"409":{"description":"A skill with the same name already exists"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]},"get":{"operationId":"findAll","summary":"Find all skills","parameters":[],"responses":{"200":{"description":"Return all skills with project and user counts"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]}},"/skills/search/{query}":{"get":{"operationId":"search","summary":"Search skills by name","parameters":[{"name":"query","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return all skills that match the query with project and user counts"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]}},"/skills/{id}":{"get":{"operationId":"findOne","summary":"Find a skill by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the skill with project and user counts"},"404":{"description":"Skill not found"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]},"patch":{"operationId":"update","summary":"Update a skill by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSkillDto"}}}},"responses":{"200":{"description":"Skill updated successfully"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Skill or category not found"},"409":{"description":"Conflict - Skill with the same name already exists"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]},"delete":{"operationId":"remove","summary":"Delete a skill by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Skill deleted successfully"},"401":{"description":"Unauthorized"},"404":{"description":"Skill not found"},"500":{"description":"Internal Server Error"}},"tags":["Skills"]}},"/skill-categories":{"post":{"operationId":"create","summary":"Create a new skill category","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSkillCategoryDto"}}}},"responses":{"200":{"description":"Skill category created successfully"},"401":{"description":"Unauthorized"},"409":{"description":"Conflict - A category with the same name already exists"},"500":{"description":"Internal Server Error - Failed to create skill category. Please try again later."}},"tags":["Skill Categories"]},"get":{"operationId":"findAll","summary":"Fetch all skill categories","parameters":[],"responses":{"200":{"description":"Return all skill categories."},"500":{"description":"Internal Server Error - Failed to find all skill categories. Please try again later."}},"tags":["Skill Categories"]}},"/skill-categories/id/{id}":{"get":{"operationId":"findOneById","summary":"Fetch a skill category by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the skill category with the specified ID."},"404":{"description":"Skill category not found"},"500":{"description":"Internal Server Error - Failed to find skill category. Please try again later."}},"tags":["Skill Categories"]}},"/skill-categories/name/{name}":{"get":{"operationId":"findOneByName","summary":"Fetch a skill category by name","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the skill category with the specified name."},"404":{"description":"Skill category not found"},"500":{"description":"Internal Server Error - Failed to find skill category. Please try again later."}},"tags":["Skill Categories"]}},"/skill-categories/{id}":{"patch":{"operationId":"update","summary":"Update a skill category by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSkillCategoryDto"}}}},"responses":{"200":{"description":"Return the updated skill category."},"401":{"description":"Unauthorized"},"404":{"description":"Skill category not found"},"409":{"description":"Conflict - A category with the same name already exists"},"500":{"description":"Internal Server Error - Failed to update skill category. Please try again later."}},"tags":["Skill Categories"]},"delete":{"operationId":"remove","summary":"Delete a skill category by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category deleted successfully."},"401":{"description":"Unauthorized"},"404":{"description":"Skill category not found"},"500":{"description":"Internal Server Error - Failed to delete skill category. Please try again later."}},"tags":["Skill Categories"]}},"/stats/summary":{"get":{"operationId":"getUserInfo","summary":"Get summary of the system's data","parameters":[],"responses":{"200":{"description":"Return summary data including total users, active users, total projects, and total organizations and more"},"500":{"description":"Internal Server Error - Failed to get summary data. Please try again later."}},"tags":["Stats"]}},"/organizations":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrganizationDto"}}}},"responses":{"201":{"description":""}}}},"/organizations/{identifier}":{"get":{"operationId":"getOrg","parameters":[],"responses":{"200":{"description":""}}}},"/organizations/{id}":{"patch":{"operationId":"updateById","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrganizationDto"}}}},"responses":{"200":{"description":""}}}},"/projects":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"getFilteredProjects","parameters":[],"responses":{"200":{"description":""}}}},"/projects/latest/{days}":{"get":{"operationId":"getLatestProjects","parameters":[{"name":"days","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/projects/in-progress/{organizationId}":{"get":{"operationId":"getInProgressProjects","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/projects/finished/{organizationId}":{"get":{"operationId":"getFinishedProjects","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/projects/{id}":{"get":{"operationId":"getProjectById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"VolunteerLink API endpoints","description":"Documentation for RESTful API endpoints of VolunteerLink project","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"},"username":{"type":"string","description":"Username of the user"},"email":{"type":"string","description":"Email address of the user"},"password":{"type":"string","description":"User password","minLength":8,"maxLength":20},"locationId":{"type":"string","description":"ID of the user's location"}},"required":["firstName","lastName","username","email","password","locationId"]},"VerifyCodeDto":{"type":"object","properties":{"code":{"type":"string","description":"Reset/Verification code sent to the user by email"},"email":{"type":"string","description":"Email address of the user"}},"required":["code","email"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","description":"User email"},"password":{"type":"string","description":"User password"}},"required":["email","password"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user for password reset"}},"required":["email"]},"UpdatePasswordDto":{"type":"object","properties":{"currentPassword":{"type":"string","description":"Current password of the user"},"password":{"type":"string","description":"The changed password"}},"required":["currentPassword","password"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user"},"newPassword":{"type":"string","description":"New password not less than 8 character","minLength":8,"maxLength":20}},"required":["email","newPassword"]},"SocialLinkDto":{"type":"object","properties":{"platform":{"type":"string","enum":["LinkedIn","GitHub","Behance","Dribbble","Instagram","Website"],"description":"Social media platform"},"url":{"type":"string","description":"URL of the social media profile"}},"required":["platform","url"]},"EducationInfoDto":{"type":"object","properties":{"field":{"type":"string","description":"Field of education"},"institute":{"type":"string","description":"Name of the institute, school, college or university"},"startDate":{"type":"string","description":"Start date of education"},"endDate":{"type":"string","description":"End date of education. If empty then ongoing education"},"description":{"type":"string","description":"Optional description of the education"}},"required":["field","institute","startDate"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"},"username":{"type":"string","description":"Username of the user"},"email":{"type":"string","description":"Email address of the user"},"bio":{"type":"string","description":"Bio text of the user"},"locationId":{"type":"string","description":"ID of the user's location"},"locationPreference":{"type":"string","description":"Location preference of the user","enum":["IN_PERSON","REMOTE","BOTH"]},"timePreference":{"type":"string","description":"Time preference of the user","enum":["SHORT_TERM","LONG_TERM","BOTH"]},"token":{"type":"string","description":"Token"},"resetCode":{"type":"string","description":"6 digit code sent to the user's email to reset password"},"verificationCode":{"type":"string","description":"6 digit code sent to the user's email to verify email"},"socialLinks":{"description":"A list of key value pairs containing the users social links","type":"array","items":{"$ref":"#/components/schemas/SocialLinkDto"}},"notificationPreference":{"description":"An object containing the users notification preference","type":"array","items":{"type":"string"}},"gender":{"type":"string","description":"Gender of the user"},"age":{"type":"number","description":"Age of the user"},"skills":{"description":"IDs of the users skills","type":"array","items":{"type":"string"}},"education":{"description":"A list of key value pairs containing the users education information","type":"array","items":{"$ref":"#/components/schemas/EducationInfoDto"}}}},"CreateLocationDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the location"},"code":{"type":"string","description":"Code of the location"}},"required":["name","code"]},"UpdateLocationDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the location"},"code":{"type":"string","description":"Code of the location"}}},"UpdateProfilePicDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user"}},"required":["email"]},"CreateSkillDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the skill"},"description":{"type":"string","description":"Description of the skill"},"categoryId":{"type":"string","description":"ID of the category to which the skill belongs"}},"required":["name","description","categoryId"]},"UpdateSkillDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the skill"},"description":{"type":"string","description":"Description of the skill"},"categoryId":{"type":"string","description":"ID of the category to which the skill belongs"}}},"CreateSkillCategoryDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the skill category"},"description":{"type":"string","description":"Description of the skill category"}},"required":["name","description"]},"UpdateSkillCategoryDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the skill category"},"description":{"type":"string","description":"Description of the skill category"}}},"CreateOrganizationDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the organization"},"mission":{"type":"string","description":"The mission statement of the organization"},"aboutUs":{"type":"string","description":"Short description of the organization"},"websiteUrl":{"type":"string","description":"The URL of the organization's website"},"locationId":{"type":"string","description":"The ID of the location where the organization is based"},"contactEmail":{"type":"string","description":"The email address of the organization's contact"},"contactPhone":{"type":"string","description":"The phone number of the organization's contact"},"foundingDate":{"format":"date-time","type":"string","description":"The founding date of the organization"}},"required":["name","mission","aboutUs","websiteUrl","locationId","contactEmail","contactPhone","foundingDate"]},"UpdateOrganizationDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the organization"},"mission":{"type":"string","description":"The mission statement of the organization"},"aboutUs":{"type":"string","description":"Short description of the organization"},"websiteUrl":{"type":"string","description":"The URL of the organization's website"},"locationId":{"type":"string","description":"The ID of the location where the organization is based"},"contactEmail":{"type":"string","description":"The email address of the organization's contact"},"contactPhone":{"type":"string","description":"The phone number of the organization's contact"},"foundingDate":{"format":"date-time","type":"string","description":"The founding date of the organization"}}},"CreateProjectDto":{"type":"object","properties":{}}}},"externalDocs":{"description":"Postman Collection","url":"/api-json"}}